requirements.txt
Django>=5.2
djangorestframework>=3.14
django-cors-headers
python-decouple
mysqlclient
celery[redis]
redis
pandas
numpy
akshare
backtrader
requests